refactor: move shared types to its file and add browser context to hooks
