+++ packages/core/src/core/directives/afterAll.ts
@@ -1,5 +1,5 @@
 import { taskRegistry } from "@/core/registry";
-import { type TaskSuiteFn } from "@/models/TaskSuite";
+import type { TaskSuiteFn } from "@/types";
 
 export function afterAll(fn: TaskSuiteFn) {
   taskRegistry.registerAfterAll(fn);
