+++ packages/core/src/core/directives/afterEach.ts
@@ -1,5 +1,5 @@
 import { taskRegistry } from "@/core/registry";
-import { type TaskCaseFn } from "@/models/TaskCase";
+import type { TaskCaseFn } from "@/types";
 
 export function afterEach(fn: TaskCaseFn) {
   taskRegistry.registerAfterEach(fn);
