+++ packages/core/src/core/directives/beforeAll.ts
@@ -1,5 +1,5 @@
 import { taskRegistry } from "@/core/registry";
-import { type TaskSuiteFn } from "@/models/TaskSuite";
+import type { TaskSuiteFn } from "@/types";
 
 export function beforeAll(fn: TaskSuiteFn) {
   taskRegistry.registerBeforeAll(fn);
