+++ packages/core/src/core/directives/beforeEach.ts
@@ -1,5 +1,5 @@
 import { taskRegistry } from "@/core/registry";
-import { type TaskCaseFn } from "@/models/TaskCase";
+import type { TaskCaseFn } from "@/types";
 
 export function beforeEach(fn: TaskCaseFn) {
   taskRegistry.registerBeforeEach(fn);
