+++ packages/core/src/core/directives/describe.ts
@@ -1,5 +1,5 @@
 import { taskRegistry } from "@/core/registry";
-import { type TaskSuiteFn } from "@/models/TaskSuite";
+import type { TaskSuiteFn } from "@/types";
 
 export function describe(name: string, fn: TaskSuiteFn) {
   taskRegistry.registerSuite(name, fn);
