+++ packages/core/src/core/directives/scenario.ts
@@ -1,5 +1,5 @@
 import { taskRegistry } from "@/core/registry";
-import { type TaskCaseFn } from "@/models/TaskCase";
+import type { TaskCaseFn } from "@/types";
 
 export async function scenario(name: string, fn: TaskCaseFn) {
   taskRegistry.registerCase(name, fn);
