+++ packages/core/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./config/schema";
 export * from "./core/directives";
 export * from "./models";
+export * from "./types";
