+++ packages/core/src/models/Renderer.ts
@@ -1,4 +1,4 @@
-import type { Trace } from "@/models/Trace";
+import type { Trace } from "@/types";
 
 export abstract class Renderer {
   public abstract render(traces: Trace[]): Promise<void>;
