--- packages/core/src/models/Trace.ts
@@ -1,5 +1,0 @@
-export interface Trace {
-  type: string;
-  data: unknown;
-  children: Trace[];
-}
