+++ packages/core/src/models/index.ts
@@ -2,4 +2,3 @@ export * from "./Renderer";
 export * from "./Task";
 export * from "./TaskCase";
 export * from "./TaskSuite";
-export * from "./Trace";
